move policy-parsing codes to utils.rs
